Add the command line parser and its tests
This patch implements the command line parser as specified in assignment
number four. This implementation assumes that the pattern and source
image arguments may be permuted from the order specified in the
assignment.

The parser throws an error whenever there are extra arguments,
insufficient arguments, or malformed arguments (i.e. a "-p" without any
subsequent filename.