Add a test-runner for all tests in the project
The project needs a file which collects all the different tests into one
file. The implementation of this is quite simple using rackunit. Each
group of tests is placed inside a `test-suite' which is `provide'd to
the test-runner. The test-runner coalesces these `test-suite's into a
larger test suite which it then runs.

The test-runner file contains a detailed description of how to add more
tests.

To execute the test-runner:

  racket run-all-tests.rkt