Add a `test' target to the Makefile
The Makefile controls our build process so it should also control our
test running process. To this end, a SPIMS_RACKET environment variable
was added, which is defined in terms of another new variable,
SPIMS_RACKET_BIN. The SPIMS_RACO variable is now also defined in terms
of SPIMS_RACKET_BIN.

The only environment variable which likely needs to be customized for
each machine is the SPIMS_RACKET_BIN variable. My racket binaries are
stored in `/usr/local/bin', so when I run make I use:

  SPIMS_RACKET_BIN=/usr/local/bin make ...

This can, of course, be packaged into an alias, a script, or you could
always set the environment variable before you start developing. I wrote
a small script, which simplifies the build process for me, with the
following content:

  #!/bin/bash
  SPIMS_RACKET_BIN=/usr/local/bin make $*

Assuming environment variables are set properly, the tests can be run by
executing:

  make test