Add data and structure definitions for pixels and images
This patch contains necessary definitions of the fundamental data types
and their associated structures. It also implements two procedures.

The `get-pixel-at' procedure is intended to abstract away the details of
the vector of vectors representation of images. `vector-ref' should not
be used on `ImageBitmap's.

The `create-bitmap' procedure abstracts away from the details of
*creating* the vector of vectors representation. The user provides a
width, height, and a generating procedure which accepts a row and column
index and produces a Pixel structure.