Add a compatibility wrapper which transforms `pixel's to `color's
Currently, we have an API mismatch in which the load-image-file module
defines the ImageBitmap data type as [Vector [Vector Pixel]] where Pixel
is a data type corresponding to the `pixel' structure. The `pixel'
structure and the `color' structure are mostly identical except that the
color structure comes from 2htdp/image. The Algorithm module expects
ImageBitmap to be a [Vector [Vector Color]]

This small wrapper converts between the two representations until we
come to a consensus on how to implement it.