Add a map and a fold for the ImageBitmap datatype in data-transformers.rkt
These procedures are necessary for the project. The tests for the
fold-bitmap procedure are essentially an implementation of a color
channel averaging that we might use as a hash code. Slightly more code
could implement color histograms.

The two procedures are:

  map-bitmap : [Pixel Number Number -> Pixel] ImageBitmap -> ImageBitmap

  fold-bitmap : FORALL(TYPE1, TYPE2).
                     [Pixel Number Number TYPE1 -> TYPE1]
                     [TYPE1 Number TYPE2 -> TYPE2]
                     TYPE1
                     TYPE2
                     ImageBitmap
                     ->
                     TYPE2

See the tests for a couple of examples of how to use the procedures.

The fold is a left fold (i.e. space efficient, but processes elements in
order like foldl [NB (foldl cons '() '(1 2 3)) => '(3 2 1) in Scheme;
in Haskell you'd get '(3 . (2 . (1 . ()))). See srfi-1 for a discussion
of why Scheme's foldl works differently from the "mathematical" foldl
that Haskell has]). *None* of that matters if you use a commutative
operation such as addition.