Add a script to run the tests associated with the file arguments
The run-all-tests.rkt file should be used in general to verify that
a patch doesn't break anything in any part of the codebase. However, it
is sometimes useful to rapidly iterate over a subset of the tests and
this command fulfills this niche.

The usage syntax is:

  ./run-test NAME.rkt ...

The test output will be written to stderr and stdout (that is Racket's
prerogative).