Add a procedure to load images that are stored in files
The load-image-file.rkt module provides the `load-image-file' procedure
which accepts a filename and produces an ImageBitmap (a datatype defined
by data-structures.rkt).

The procedure relies on the racket/draw library to load the image. The
racket/draw library supports a superset of the image formats that the
project requirements demand. The `load-image-file' procedure does not
signal an error when given a file which is supported by racket/draw but
not listed in the project requirements.

In the case of a non existent file, it signals an error.

In the case of a file which does not contain an image in one of
racket/draw's valid formats, it signals an error.