Add tests for the image loading procedure in `load-image-file.rkt'
These tests load images and compare them to programmatically generated
`ImageBitmap's. These tests depend on a series of simple image files
that are stored in tests/images/. These files are included in this patch
because any changes to them affect the functionality of the tests and
are necessary for said tests to successfully execute.

Additionally, this patch modifies the `test' target of the Makefile in a
backwards compatible way. Previously, the working directory of the
Makefile was irrelevant to the running of the tests. Now the working
directory is changed to the tests/ folder for the duration of the
tests. This is the simplest way to ensure the relative paths in
`load-image-file.rkt' are correctly resolved.