Add implementation and tests for parse-command-line with directories
Because parse-command-line decodes directories into lists of paths, the
file/directory existence checks now reside in parse-command-line and not in
load-image-file.

Unfortunately, Racket's `command-line' macro does not support flags which have
one dash and multiple characters. As a result, there is a hack called "fix-args"
which converts the external command line syntax to a syntax supported by the
`command-line' macro.