Fix all type errors in data-structures.rkt
Change uses of `(Vector X)' to the intended `(Vectorof X)'. Change uses
of `Number' to `Real' or `Integer'.

(Vector X Y Z) means a vector of exactly length three whose first
element is of type X, second element is of type Y, and third element is
of type Z.

`Number' is the union of `Real' and `Complex'. Don't use `Number' unless
you're absolutely certain that you want your input to possibly be a
complex number. Instead use `Real'.

When in doubt use `Real'. If that doesn't work try `Integer'.