Differentiate between fast and slow test suites
To avoid running the slow tests (i.e. all the A5/A4 tests) just run:

  racket run-all-tests.rkt fast

To run the full test suite, drop "fast". I've also added this to the
Makefile, so we can now run

  make fast-test

to skip the slow tests, we can still run

  make test

to run the full test suite.